file to interfaces, fix build error
